feat: User에 있던 emailVerified를 Member만 사용하도록 변경
