+++ src/main/java/com/sky7th/deliveryfood/security/provider/AbstractAuthenticationProvider.java
@@ -1,6 +1,5 @@
 package com.sky7th.deliveryfood.security.provider;
 
-import com.sky7th.deliveryfood.security.exception.EmailNotVerifiedException;
 import com.sky7th.deliveryfood.user.CustomUserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationProvider;
@@ -32,9 +31,6 @@ public abstract class AbstractAuthenticationProvider implements AuthenticationPr
 
     } else if (!userDetails.isAccountNonLocked()) {
       throw new LockedException(email);
-
-    } else if (!userDetails.isEnabled()) {
-      throw new EmailNotVerifiedException(email);
     }
 
     userDetails.setPassword(null);
