+++ src/main/java/com/sky7th/deliveryfood/user/CustomUserDetails.java
@@ -46,7 +46,7 @@ public class CustomUserDetails extends User implements UserDetails {
 
     @Override
     public boolean isEnabled() {
-        return super.getEmailVerified();
+        return true;
     }
 
     @Override
